requirements.txt
PyYAML>=6.0,<7.0
typer>=0.9.0,<0.10.0
